extract shared JSONBig parser into json.ts
- Create shared json.ts module with BigInt-safe JSON parsing
- Use storeAsString: false with custom reviver for proper BigNumber handling
- Remove duplicated code from client.ts and config.ts
