+++ templates/cli/lib/client.ts
@@ -1,9 +1,8 @@
 import os from "os";
 import { fetch, FormData, Agent } from "undici";
-import JSONbigModule from 'json-bigint';
-import { BigNumber } from 'bignumber.js';
 import { AppwriteException } from "@appwrite.io/console";
 import { globalConfig } from "./config.js";
+import { JSONBig } from "./json.js";
 import chalk from "chalk";
 import type {
   Headers,
@@ -19,32 +18,6 @@ import {
   SDK_VERSION,
   SDK_TITLE,
 } from "./constants.js";
-
-const JSONbigParser = JSONbigModule({ storeAsString: false });
-const JSONbigSerializer = JSONbigModule({ useNativeBigInt: true });
-
-const MAX_SAFE = BigInt(Number.MAX_SAFE_INTEGER);
-const MIN_SAFE = BigInt(Number.MIN_SAFE_INTEGER);
-
-function reviver(_key: string, value: any): any {
-  if (BigNumber.isBigNumber(value)) {
-    if (value.isInteger()) {
-      const str = value.toFixed();
-      const bi = BigInt(str);
-      if (bi >= MIN_SAFE && bi <= MAX_SAFE) {
-        return Number(str);
-      }
-      return bi;
-    }
-    return value.toNumber();
-  }
-  return value;
-}
-
-const JSONbig = {
-  parse: (text: string) => JSONbigParser.parse(text, reviver),
-  stringify: JSONbigSerializer.stringify
-};
 
 class Client {
   private endpoint: string;
@@ -224,7 +197,7 @@ class Client {
 
       body = formData;
     } else {
-      body = JSONbig.stringify(params);
+      body = JSONBig.stringify(params);
     }
 
     let response: Awaited<ReturnType<typeof fetch>> | undefined = undefined;
@@ -291,7 +264,7 @@ class Client {
     const text = await response.text();
     let json: T | undefined = undefined;
     try {
-      json = JSONbig.parse(text);
+      json = JSONBig.parse(text);
     } catch (error) {
       return text as T;
     }
